requirements.txt
camelot-py==0.11.0
opencv-python==4.8.0.76
